requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
Django==3.2.8
django-crispy-forms==1.13.0
django-filter==21.1
future==0.18.2
html5lib==1.1
Pillow==8.4.0
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.3
reportlab==3.6.2
six==1.16.0
sqlparse==0.4.2
webencodings==0.5.1
xhtml2pdf==0.2.5
